+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## UNRELEASED
 
+## Fixes
+
+* Fixed missing translation for "New Piece" option on the "more" menu of the piece manager, seen when using it as a chooser.
+* Piece types with relationships to multiple other piece types may now be configured in any order, relative to the other piece types. This sometimes appeared to be a bug in reverse relationships.
+* Code at the project level now overrides code found in modules that use `improve` for the same module name. For example, options set by the `@apostrophecms/seo-global` improvement that ships with `@apostrophecms/seo` can now be overridden at project level by `/modules/@apostrophecms/global/index.js` in the way one would expect.
+
+## 3.5.0 - 2021-09-23
+
 ### Fixes
 
 * Pinned dependency on `vue-material-design-icons` to fix `apos-build.js` build error in production.
@@ -13,7 +21,7 @@
 * `css-loader` now ignores `url()` in css files inside `assets` so that paths are left intact, i.e. `url(/images/file.svg)` will now find a static file at `/public/images/file.svg` (static assets in `/public` are served by `express.static`). Thanks to Matic Tersek.
 * Restored support for clicking on a "foreign" area, i.e. an area displayed on the page whose content comes from a piece, in order to edit it in an appropriate way.
 * Apostrophe module aliases and the data attached to them are now visible immediately to `ui/src/index.js` JavaScript code, i.e. you can write `apos.alias` where `alias` matches the `alias` option configured for that module. Previously one had to write `apos.modules['module-name']` or wait until next tick. However, note that most modules do not push any data to the browser when a user is not logged in. You can do so in a custom module by calling `self.enableBrowserData('public')` from `init` and implementing or extending the `getBrowserData(req)` method (note that page, piece and widget types already have one, so it is important to extend in those cases).
-* Code at the project level now overrides code found in modules that use `improve` for the same module name. For example, options set by the `@apostrophecms/seo-global` improvement that ships with `@apostrophecms/seo` can now be overridden at project level by `/modules/@apostrophecms/global/index.js` in the way one would expect.
+* `options.testModule` works properly when implementing unit tests for an npm module that is namespaced.
 
 ### Changes
 
@@ -21,7 +29,6 @@
 
 ### Adds
 
-* If `options.testModule` on the app is a string it will be used as an npm namespace when creating a symlink test module.
 * Rich Text widget's styles support a `def` property for specifying the default style the editor should instantiate with.
 * A more helpful error message if a field of type `area` is missing its `options` property.